==> 3_getIntraInterPlots.R <==
## Script to get MI boxplots and barplots for the different interaction types:
## Inter-Cytoband, Intra-Cytoband, Trans per condition.
## It requires the files generated by 2_cutNetworks

library(dplyr)
library(ggplot2)
library(ggthemes)

n <- 20127  
conds <- c("healthy", "luma")
colors <- c("#e3a098", "#a32e27")
labels <- c("Healthy", "Luminal A")
names(colors) <- labels

all_interactions <- lapply(X = conds, FUN = function(cond) {
  cat("Reading network ", cond, "\n")
  interactions <- read_tsv(paste0("data/network-tables/", cond, "-", n, "-interactions.tsv"))
  interactions$cond <- cond
  return(interactions)
})

all_interactions <- bind_rows(all_interactions)
all_interactions$interaction_type <- factor(all_interactions$interaction_type, 
                                            levels = c("Intra-Cytoband", "Inter-Cytoband", "Trans"))
all_interactions$cond <- factor(all_interactions$cond,  levels = conds, labels = labels)

p <- ggplot(all_interactions) + 
  geom_boxplot(aes(x = interaction_type, y = MI,  fill = interaction_type)) +
  theme_few(base_size = 20) +
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text.x = element_text(size = 16),
        axis.title.x = element_blank()) + 
  facet_wrap(~cond, nrow = 1) + 
  ylab("MI") +
  scale_fill_manual(name = "Interaction", values =  c("#1696AC", "#73299e", "#C85200")) 

png(paste0("figures/mi-boxplot-by-inter-type.png"), width = 1000, height = 500)
plot(p)
dev.off()  

p <- ggplot(all_interactions, aes(x = interaction_type, fill = interaction_type)) + 
  geom_bar() +
  geom_text(stat='count', aes(label=..count..), vjust=-1) +
  ylim(c(0,20000)) +
  theme_few(base_size = 20) +
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text.x = element_text(size = 16),
        axis.title.x = element_blank()) + 
  facet_wrap(~cond, nrow = 1) + 
  ylab("Count") +
  scale_fill_manual(name = "Interaction", values =  c("#1696AC", "#73299e", "#C85200")) 

png(paste0("figures/mi-barplot-by-inter-type.png"), width = 1000, height = 500)
plot(p)
dev.off()
